requirements.txt
glfw==2.5.4
keyboard==0.13.5
Nuitka==1.0.5
numpy==1.23.2
Pillow==9.2.0
PyOpenGL==3.1.6
PySide6==6.3.1
PySide6-Addons==6.3.1
PySide6-Essentials==6.3.1
shiboken6==6.3.1
